Remove the special nature of the publish-utils name property. (#30)
* Refactor PublishTasks to PublishPlugin.
* Remove the "special" nature of publish-utils' name property.
* Update changelog.
