+++ changelog.md
@@ -2,6 +2,38 @@
 
 ## Version 4
 
+### Version 4.0.7
+
+* `publish-utils`: Setting the `name` property no longer triggers configuration. This is now applied after the project has been evaluated.
+* `publish-utils`: Add `publishJavadoc` property so that publishing Javadoc can be optional.
+
+### Version 4.0.6
+
+* `api-scanner`: Field annotations are now written out in alphanumerical order. The `@JvmField` annotation is no longer written at all.
+
+### Version 4.0.5
+
+* `cordformation`: Fix webserver startup when headless.
+
+### Version 4.0.4
+
+* `api-scanner`: Class/interface/method annotations are now written out in alphanumerical order.
+* `api-scanner`: Handle methods with vararg parameters correctly.
+
+### Version 4.0.3
+
+* `cordformation`: Fix parsing of unknown node configuration properties.
+
+### Version 4.0.2
+
+* `cordformation`: Add `NetworkParameters` contract implementation whitelist.
+
+### Version 4.0.1
+
+* `cordformation`: Use updated rpc users config format.
+
+### Version 4.0.0
+
 * Split repository from [Corda](https://github.com/corda/corda) repository.
 
 ## Pre-version 4
